Filter `codesign` messages
Right now we end up with this output on relink:

```
/Users/neonacho/Projects/public/swift-package-manager/.build/arm64-apple-macosx/debug/swift-bootstrap: replacing existing signature
```

This change specifically filters any output from `codesign` commands which haven't failed and we aren't in verbose mode.

This should eventually be possible at the command level, but for the time being this allows us to avoid shipping SwiftPM with this messy output.

rdar://118937939