command plugins: Test dependencies are always built for the host
This integration test checks that any targets depended on by a
command plugin are built for the host, not for the target.

* A new CommandPluginTestStub plugin has a dependency on a target
  executable which will be built automatically when the plugin is
  run.   The test checks that the dependency is built for the host
  architecture, no matter which target architecture is selected
  using '--triple'.

* The plugin also asks SwiftPM to build the 'placeholder' main
  target.   The test checks that the dependency is built for the
  target architecture.

The test is restricted to macOS because we can be sure of having
a viable cross-compilation environment (arm64 to x86_64 and vice
versa).  The standard Linux build environments can't cross compile
to other architectures.