Add mermaid charts support to `swift package describe` (#7289)
This adds support for new `--type mermaid` option on `swift package describe`.

This format is supported by the GitHub Markdown renderer and some Markdown preview extensions in IDEs.

I'm using different shapes to denote difference between products and targets. `| Foo |` is a target, `|| Bar ||` is a product, and hexagonal shapes declared with `{{ Baz }}` are products from package dependencies. I initially tried to include a legend subgraph, but that makes layout quite ugly even for simple packages, so I decided to exclude that by default. The codepath for rendering the legend is still there, so we can reenable it at any point or make conditional on some new CLI flag if needed.