Inherit host Swift SDK's toolset paths to fallback to host tools
Currently, Swift SDK fallbacks to system `PATH` search path for tools
that are not found in the SDK. Usually, host compiler tool is capable to
cross compile but just lacks some of artifacts like platform libraries.
In that case, we don't need to have a separate compiler tools in
target-specific Swift SDK, but we can share the host compiler tools as
long as the host Swift and target Swift are the same version.

This patch makes Swift SDK fallback to host Swift SDK tools if the
target Swift SDK doesn't have such tools. This unlocks eliminating
compiler tools from target-specific Swift SDKs.