command plugins: Optionally print build logs during packageManager.build (#7255)
This commit adds the option for a build run on behalf of a command
plugin to echo logs to the console as they are produced, in addition to
the current behaviour of accumulating them in a buffer. Console echoing
is turned off by default.

### Motivation:

When a plugin asks for a target to be built by calling
`packageManager.build`, SwiftPM accumulates the build log in a buffer
and returns the whole log when the build completes. No feedback is
printed on the console while the build is running, so if the target is
large the user can't tell whether the build is making progress or is
stuck.

### Modifications:

* Added `TeeOutputByteStream`, an `OutputByteStream` which copies its
inputs to a set of output streams
* Added an `echoLogs` parameter to `packageManager.build`, defaulting to
`false` to match the current behaviour.
* Added a test to verify that logs are echoed and the existing log
accumulation continues to work.
 
### Result:

A command plugin can ask for build outputs to be echoed as they are
produced. The user will have some feedback that a build is happening,
rather than the long pause which can happen now.

This change depends on #7254.