Fix non-tarballed SDK installation with remote URL
4714ea92a introduced a regression where non-tarball SDKs could not be
installed from a remote URL due to the wrong assumption that the
downloaded file would always be a tarball. This commit fixes the issue
by checking the file extension part of the URL for every supported
archive format, then falling back to the tarball format if no supported
extension is found.