Make `.dir-locals.el` sensible
The variables set in `swiftpm-project-settings.el` can be set directly
in `dir-locals.el`. There is no reason to move them to a separate file,
and in fact doing that will automatically make Emacs flag the dir-locals
as unsafe, since it will be being requested to evaluate a block of code,
instead of setting known-safe variables.

Signed-off-by: Julia DeMille <me@jdemille.com>