Move `swiftpm-project-settings.el` settings to `.dir-locals.el` (#7314)
Directory-local variables have been moved so that they are set directly
in `.dir-locals.el`.

### Motivation:

The variables set in `swiftpm-project-settings.el` can be set directly
in `dir-locals.el`. There is no reason to move them to a separate file,
and in fact doing that will automatically make Emacs flag the dir-locals
as unsafe, since it will be being requested to evaluate a block of code,
instead of setting known-safe variables.

### Modifications:

`Utilities/Emacs/swiftpm-project-settings.el` has been removed. All
variables
set in it have been moved to `.dir-locals.el` itself. This allows the
evaluation
of all directory locals to be considered safe.

Additionally, `swift-basic-offset` was changed to
`swift-mode:basic-offset`.
That appears to be the proper variable for setting indent width for
Swift in Emacs.

### Result:

Emacs users can edit the project more peacefully.

Signed-off-by: Julia DeMille <me@jdemille.com>