Allow empty targets if plugins are applied to them
Currently, we warn about and remove empty targets in `PackageBuilder` which can be unexpected if the sole purpose of a target is to host generated sources or resources from plugins. This change makes it so that we treat these targets as Swit targets and potentially remove them later on during the build.

rdar://92858144