Rename `Swift*Tool` to `Swift*Command`
This fixes an existing point of confusion, where `SwiftBuildTool` that handles `swift build` CLI invocations could be misunderstood as something responsible for build tools in plugins, which becomes especially apparent when working on the plugins codebase.

Additionally, `SwiftTool` is a grab bag reference type for storing state injected into a CLI command, so we should rename it to `SwiftCommandState` for clarity.