Add option to allow HTTP (not HTTPS) registries (#7204)
Addresses #7170: this adds an option to the `package-registry` `set` and
`publish` commands to allow using an HTTP registry endpoint, bypassing
the existing requirement that package registries use HTTPS.

### Motivation:

In short, the motivation is to allow publishing to a registry that is
hosted on a local machine. It is difficult to host a fully TLS-compliant
HTTPS endpoint on developer machines at scale. See more detail in #7170

### Modifications:

* Adds a new `allowInsecureHTTP` flag to the Registry model.
* Adds a new `--allow-insecure-http` flag to the `package-registry`
`set` and `publish` commands.
* Adds tests that validate the behavior of the model and command
additions.
* Adds tests that validate the requested constraints in #7170 regarding
HTTP endpoints, such as not being able to publish to an HTTP endpoint if
there is authentication configured for it, and `login` command requiring
HTTPS.

### Result:

We will be able to use the `swift` CLI to publish to our HTTP registries
locally hosted on our development machines. As a result, we will be able
to migrate to the official tooling instead of needing to continue using
home-rolled publish tooling.