Support for prebuilt packages in the SDK (#7337)
This should allow satisfying a package dependency with a package in the
SDK (or toolchain) based on metadata. If the given prebuilt library is
version-compatible, we essentially elide the dependency from the graph.
If it is not, we will fetch and build the dependency as normal. If a
library with associated metadata is linked without a corresponding
package dependency, we'll emit a warning.