Support macros when cross-compiling (#7118)
### Motivation:

Not supporting macros in cross-compilation is a major limitation,
especially for libraries like https://github.com/apple/swift-mmio.

### Modifications:

Added `enum BuildTriple { case tools, destination }` and `var
buildTriple: BuildTriple` on `ResolvedTarget` and `ResolvedProduct`.
We're not using "host" and "target" triple terminology in this enum, as
that clashes with build system "targets" and can lead to confusion in
this context.

Corresponding value is assigned to this property depending on target and
product type: `tools` for macros, plugins, and their dependencies,
`destination` for everything else (the default). Based on this property
we can choose between `buildParameters.hostTriple` and
`buildParameters.targetTriple` during build plan construction.

Additionally, the resolved products and resolved targets graph is now
constructed in a way that allows certain targets and products to be
built twice: once for host triple, once for target triple if needed.
This required modifying build description and build manifest generation
to distinguish these products and targets from each other that are built
twice.

Artifacts built for the host now have `-tools` suffix appended to their
names. This cascaded into making changes throughout the code base for
build tool plugins and package plugins handling, which constructed their
paths in an ad-hoc manner without accounting for possible changes to
these names.

Also added `CrossCompilationPackageGraphTests` and
`CrossCompilationBuildPlanTests` to verify the changes made are applied
correctly.

### Result:

Resolves https://github.com/apple/swift-package-manager/issues/6950
Resolves rdar://105991372