Build: pass through `Embedded` flag to link jobs (#7304)
### Motivation:

Without passing this flag, Swift Driver links such products as if
they're not built for embedded platforms at all, passing incorrect flags
to the linker. This is not reproducible when using plain `swiftc`, as
that combines both compile and link jobs by default, but SwiftPM prefers
to run those separately, which requires this special handling.

### Modifications:

Directly checking in `ProductBuildDescription/linkArguments` whether all
of the product's targets are built in the embedded mode. If so, we're
passing the embedded mode flag to Swift Driver.

Unfortunately, `BuildSettings.AssignmentTable` is formed too early in
the build process right in `PackageModel`, while we still need to run
checks specific build settings quite late in the build stage. Because of
that, there's no clean way to check if `Embedded` flag is passed other
than directly rendering `BuildSettings.Scope` to a string and running a
substring check on that. In the future we should move `BuildSettings`
out of `PackageModel`, ensuring that SwiftPM clients don't rely on this
behavior anymore.

### Result:

Products that have targets using Embedded Swift can be built with
SwiftPM, assuming that Swift Driver handles other linker flags
correctly.