Rename `PackageGraph` type to `ModulesGraph`
We have to operate on two different graphs in the `PackageGraph` module:
1. A graph of packages in PubGrub package resolution code;
2. A graph of resolved products and modules that belong to those products.

Currently the second graph is misleadingly called `PackageGraph`, although that name is much more suitable for the first graph. This naming is confusing and unfortunate, and can be especially misleading for first-time contributors. We should better document the SwiftPM resolution and build pipeline in the future, but cleaning up naming is the first step.

In the short term we should also split the `PackageGraph` module into `PubGrub` and `ModulesGraph` modules, but for now I'm renaming a single type to keep the PR manageable.