Rename `Swift*Tool` to `Swift*Command` (#7336)
This removes an existing point of confusion, where `SwiftBuildTool` that
handles `swift build` CLI invocations could be misunderstood as
something responsible for build tools in plugins. It becomes especially
problematic when working on the plugins codebase.

Additionally, `SwiftTool` is a grab bag reference type for storing state
injected into a CLI command, so we should rename it to
`SwiftCommandState` for clarity.