Allow `swift-build` to have `async` entrypoint (#7369)
Due to a bug in Swift 5.9, we can't have multiple executables with
`async` entrypoints linked into test products. This was the reason why
we weren't able to make `swift-build` entrypoint `async`. Since we have
a single `swift-package-manager` executable anyway, we can link that one
into tests instead, and pass "simulated" executable name via an
environment variable. This allows us to work around the bug and start
using structured concurrency from `swift-build` entrypoint.

This change is a prerequisite for
https://github.com/apple/swift-package-manager/pull/7367.