Benchmarks: add `PackageGraphBenchmarks` target (#7283)
This adds a dependency on the
[`package-benchmark`](https://github.com/ordo-one/package-benchmark)
library, isolated to the new `Benchmarks` folder. Thus it doesn't touch
CMake or self-hosted builds, and one can only pull the library if they
run `swift package benchmark` in the `Benchmarks` directory.

A single benchmark is added that loads SwiftPM's `Package.swift` and
builds a package graph for it.