Add swift-testing support to `swift build --build-tests`.
This PR generalizes the enable/disable XCTest/swift-testing options from
`swift test` and `swift package init` and adds them to `swift build --build-tests`
so that it is possible to build all test content without actually running the
tests. If neither flag is specified, the default test content is built (always
XCTest, and swift-testing if it's a dependency, same as `swift test`.)

I don't have unit tests for this work because SwiftPM cannot currently reference
or link to swift-testing in CI. I'm open to ideas here.

Resolves #7375.