Fix a memory leak in `DownloadTaskManager` and `DataTaskManager` (#7408)
`DownloadTaskManager` had a strong reference to `URLSession`, which
retained its delegate, which was the `DownloadTaskManager`. This formed
a retain cycle.

Make the reference from `URLSession` to the delegate weak and only keep
`DownloadTaskManager` alive as long as `DownloadTask`s need it. This
causes the `DownloadTaskManager` to be deallocated once nobody holds a
reference to it anymore and all the tasks it manages are finished.

Finally, we need to call `invalidate` on the `URLSession` to tell it
that we’re done and that it should release its delegate (which is now
the `WeakDownloadTaskManager`wrapper).

This retain cycle was causing a leak in sourcekit-lsp. I verified that
the leak no longer exists with this patch.

rdar://125012584