[PubGrub] Avoid resolving package versions twice in presence of prebuilt libraries
Integrate information about versions of prebuilt libraries into the solver
and prioritize these versions when making decisions if they meet all of the
requirements of undecided terms.

This helps solver to avoid fetching and at the same time allows partual
matches on prebuilt libraries with fallback to package provided versions.