Add basic `QueryEngine` scaffolding w/ `Encodable` hashing (#7347)
### Motivation:

Caching of computations in SwiftPM is ad-hoc and mostly relies on
in-memory `ThreadSafeKeyValueStore`, which utilizes locking and makes it
harder to add `Sendable` conformances on types that have these stores as
properties. To allow consistent and persistent caching when SwiftPM
processes are relaunched, we can use a SQLite-backed async-first caching
engine.

### Modifications:

This change ports [most of the current `GeneratorEngine`
implementation](https://github.com/apple/swift-sdk-generator/blob/main/Sources/GeneratorEngine/Engine.swift)
from the Swift SDK Generator repository to the SwiftPM code base as
`QueryEngine`. Since SwiftNIO is not supported on Windows, references to
`AsyncHTTPClient` have been removed. Additionally, we can't use macros
in the SwiftPM code base either, thus the `Query` protocol has to
conform to `Encodable` instead of using macro-generated conformances. We
don't have a consistent hashing implementation for `Encodable` yet, and
a temporary stub for it is marked with `fatalError` for now.

### Result:

NFC, new `QueryEngine` module is not used anywhere yet.