Add a PackageModelSyntax library that manipulates the source of a manifest file
Package manifest files are Swift source code, so editing them means working
with the source code directly. Introduce a new library based on swift-syntax
that allows us to perform targeted manipulations of the manifest file
programmatically, making suggested edits to the file without having to
understand everything in it.

Introduce one editing operation that adds a particular package dependency
to the manifest file. For example, it takes the programmatic representation
of a package dependency in the model
(`PackageModel.PackageDependency`) and the syntax for a manifest file,
then it will produce a set of edits that extend (or add) the
dependencies in the `Package` instance to the manifest file, e.g.,

    .package(url: "https://github.com/apple/swift-syntax.git", from: "510.0.1"),

We make an attempt to match the surrounding trivia so that we don't make
an ugly mess of the resulting manifest. I have tests for a number of
cases to make sure they look nice, but I expect we'll have to refine
the heuristics over time.