Add simple synthetic modules graph benchmarks (#7465)
New `SyntheticModulesGraph` benchmark is introduced, which calls
`loadModulesGraph` function that we already run against test fixtures in
`ModulesGraphTests` and `BuildPlanTests`. The function under benchmark
almost immediately delegates to `ModulesGraph.load`, which is used in
real-world modules graph dependency resolution.

Benchmark parameters are controlled with
`SWIFTPM_BENCHMARK_MODULES_GRAPH_DEPTH` and
`SWIFTPM_BENCHMARK_MODULES_GRAPH_WIDTH` environment variables, so in the
future we should be able to plot graphs of benchmark metrics against
depth and width of modules graph over a given range.

Thresholds are now split into platform-specific directories so that
thresholds recorded on x86_64 don't interfere with thresholds for arm64,
same for the OS family.