Render package dependencies with the SwiftPM 5.5 syntax
Modern SwiftPM implementations warn about the

  .package(url: "https://github.com/apple/swift-system.git", .branch("main"))

syntax that's being rendered into the package manifest, because this syntax
was replaced in SwiftPM 5.5 with something that uses argument labels:

  .package(url: "https://github.com/apple/swift-system.git", branch: "main")

Since generating code with warnings is bad, and we've already stated
that we won't edit packages for tools version < 5.5, switch to this new
style.

This does mean that we drop the (internal) dependency on
ManifestSourceGeneration, which is producing the older format, because
I don't have a strong sense of whether we can modernize that code.
Instead, we have a new path that generates swift-syntax nodes directly
for the types we generate.