[SE-0301] Introduce `swift package add-target` support to add targets to a package
Add a new package command to add a target with the given name, type, and
dependencies to the package. This includes adding the target to the
package manifest as well as creating stub source files for the target to
guide the user. For example, this command:

    swift package add-target SocialGraphClientTests --dependencies SocialGraphClient --type test

adds the following to the targets in the package manifest:

    .testTarget(name: "SocialGraphClientTests", dependencies: ["SocialGraphClient"]),

as well as creating the file `Tests/SocialGraphClientTests/SocialGraphClientTests.swift`,
which looks like this:

    import SocialGraphClient
    import XCTest

    class SocialGraphClientTests: XCTestCase {
        func testSocialGraphClientTests() {
            XCTAssertEqual(42, 17 + 25)
        }
    }

There is, undoubtedly, some tuning to do to clean this up. Here's the
command-line interface, which mostly aligns with SE-0301:

    OVERVIEW: Add a new target to the manifest

    USAGE: swift package add-target <name> [--type <type>]
[--dependencies <dependencies> ...] [--url <url>] [--path <path>]
[--checksum <checksum>]

    ARGUMENTS:
      <name>                  The name of the new target

    OPTIONS:
      --type <type>           The type of target to add, which can be
one of
                              (default: library)
      --dependencies <dependencies>
                              A list of target dependency names
      --url <url>             The URL for a remote binary target
      --path <path>           The path to a local binary target
      --checksum <checksum>   The checksum for a remote binary target
      --version               Show the version.
      -h, -help, --help       Show help information.