build: remove the requirement to checkout swift-syntax (#7486)
In the case that a shared copy of swift-syntax is being used to build,
allow the sources to be unavailable. If the user specifies
`SWIFTPM_PATH_TO_SWIFT_SYNTAX_SOURCE`, that will be preferred,
otherwise, we fallback to the upstream git repository.

(cherry picked from commit dbd894521ad595036c2912bb056ebd4b6a3336c3)