[6.0] plugins: Pass correct dependency origin information to plugins
Cherry pick of https://github.com/apple/swift-package-manager/pull/7506

SwiftPM sends information about the origin of each package dependency
in the plugin context, but the PluginContext passed to the plugin
function incorrectly showed all origins as 'root'.   This happened
because the origin field was hard-coded to 'root' in the plugin-side
deserialization function.