Adds prepare for index experimental build argument.
This will be used by sourcekit-lsp to build the swiftmodule
files it needs for indexing.

Adds experimental-prepare-for-indexing flag to swift build.
Creates a llbuild manifest specific for the prepare build that
skips generating object files and linking of those files and
calls swiftc to only create the swiftmodule as quickly as it can.