[Package/ModuleGraph] Allow cyclic package dependencies if they don't introduce a cycle in a build graph (#7530)
It should be possible for packages to depend on each other if such
dependence doesn't introduce cycles in the build graph.

- Introduced a new DFS method to walk over graphs that breaks cycles.
- Replaces use of `findCycle` + `topologicalSort` with `DFS` while
building manifest and package graphs. This allows cycles in dependencies
to be modeled correctly.
- Removes some of the redundant data transformations from modules graph.
- Modifies `ResolvedPackage` to carry identities of its dependencies
instead of resolved dependencies themselves. This helps to simplify
logic in `createResolvedPackages`.
- Adds detection of target cycles across packages.

Makes it possible for package A to depend on package B and B to depend
on A if their targets don't form a cycle.

(cherry picked from commit 8b129093e02bb30c2dd9d3af5187435aba25f53e)