plugins: Global compiler flags should not be used when building plugin dependencies
Build tools and end products are built with separate toolchains:
* Plugins and any targets they depend on, such as helper tools, are
built with the host toolchain.
* End products are built with the target toolchain.

Compiler flags set with -Xswiftc are intended to be applied only
to end product builds, but they are incorrectly being applied to
plugin tool builds as well.   This can cause problems because target
toolchain flags might not be suitable for a tools build.   In the
worst case, the build might fail completely if the host toolchain
rejects the target toolchain flag.    For example, Linux toochains
accept the `-static-executable` flag, but macOS toolchains reject
it.

Fixes: rdar://127813618